Proyecto-ventas/apps/usuarios/views.py
from rest_framework import viewsets
from .models import Usuario
from .serializer import UsuarioSerializer
from .permissions import EsAdministrador
from rest_framework.authentication import SessionAuthentication, BasicAuthentication
from rest_framework.permissions import IsAuthenticated
 
 
class UsuarioViewSet(viewsets.ModelViewSet):
    queryset = Usuario.objects.all()
    serializer_class = UsuarioSerializer
    authentication_classes = [SessionAuthentication, BasicAuthentication]
    permission_classes = [IsAuthenticated, EsAdministrador]