requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
djangorestframework==3.16.1
Markdown==3.10.2
sqlparse==0.5.5
tzdata==2025.3
drf-yasg==1.21.15
